jdk升级为21，springboot升级为3.3.5，maven升级为3.9.9

+++ src/main/java/com/itheima/reggie/service/impl/DishServiceImpl.java
@@ -51,7 +51,7 @@ public class DishServiceImpl extends ServiceImpl<DishMapper, Dish> implements Di
     public int getCountByCategoryId(Long categoryId) {
         LambdaQueryWrapper<Dish> dishWrapper = Wrappers.lambdaQuery();
         dishWrapper.eq(Dish::getCategoryId, categoryId);
-        return this.count(dishWrapper);
+        return (int) this.count(dishWrapper);
     }
 
     /**
